Collation: Fixed "Bitwise-or operator used on a sign-extended operand; consider casting to a smaller unsigned type first" warnings

+++ src/ICU4N/ICU4N.Collation/Impl/Coll/BOCSU.cs
@@ -250,7 +250,7 @@ namespace ICU4N.Impl.Coll
                 --result;
                 modulo += factor;
             }
-            return (result << 32) | modulo;
+            return (result << 32) | (uint)modulo;
         }
 
         /// <summary>
